+++ packages/share_plus/share_plus/android/src/main/kotlin/dev/fluttercommunity/plus/share/MethodCallHandler.kt
@@ -1,5 +1,6 @@
 package dev.fluttercommunity.plus.share
 
+import android.os.Build
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import java.io.IOException
@@ -11,48 +12,34 @@ internal class MethodCallHandler(
 ) : MethodChannel.MethodCallHandler {
 
     override fun onMethodCall(call: MethodCall, result: MethodChannel.Result) {
+        // The user used a *WithResult method
+        val isResultRequested = call.method.endsWith("WithResult")
+        // We don't attempt to return a result if the current API version doesn't support it
+        val isWithResult = isResultRequested && Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP_MR1
+
         when (call.method) {
-            "share" -> {
+            "share", "shareWithResult" -> {
                 expectMapArguments(call)
+                if (isWithResult && !manager.setCallback(result)) return
+
                 // Android does not support showing the share sheet at a particular point on screen.
                 share.share(
                     call.argument<Any>("text") as String,
                     call.argument<Any>("subject") as String?,
-                    false,
+                    isWithResult,
                 )
-                result.success(null)
-            }
-            "shareFiles" -> {
-                expectMapArguments(call)
 
-                // Android does not support showing the share sheet at a particular point on screen.
-                try {
-                    share.shareFiles(
-                        call.argument<List<String>>("paths")!!,
-                        call.argument<List<String>?>("mimeTypes"),
-                        call.argument<String?>("text"),
-                        call.argument<String?>("subject"),
-                        false,
-                    )
-                    result.success(null)
-                } catch (e: IOException) {
-                    result.error("Share failed", e.message, null)
+                if (!isWithResult) {
+                    if (isResultRequested) {
+                        result.success("dev.fluttercommunity.plus/share/unavailable")
+                    } else {
+                        result.success(null)
+                    }
                 }
             }
-            "shareWithResult" -> {
+            "shareFiles", "shareFilesWithResult" -> {
                 expectMapArguments(call)
-                if (!manager.setCallback(result)) return
-
-                // Android does not support showing the share sheet at a particular point on screen.
-                share.share(
-                    call.argument<Any>("text") as String,
-                    call.argument<Any>("subject") as String?,
-                    true,
-                )
-            }
-            "shareFilesWithResult" -> {
-                expectMapArguments(call)
-                if (!manager.setCallback(result)) return
+                if (isWithResult && !manager.setCallback(result)) return
 
                 // Android does not support showing the share sheet at a particular point on screen.
                 try {
@@ -61,8 +48,16 @@ internal class MethodCallHandler(
                         call.argument<List<String>?>("mimeTypes"),
                         call.argument<String?>("text"),
                         call.argument<String?>("subject"),
-                        true,
+                        isWithResult,
                     )
+
+                    if (!isWithResult) {
+                        if (isResultRequested) {
+                            result.success("dev.fluttercommunity.plus/share/unavailable")
+                        } else {
+                            result.success(null)
+                        }
+                    }
                 } catch (e: IOException) {
                     result.error("Share failed", e.message, null)
                 }